feat: abstract achievements
