+++ package.json
@@ -23,6 +23,7 @@
     "react-loader-spinner": "^4.0.0",
     "react-router-dom": "^6.2.1",
     "react-scripts": "^5.0.1",
+    "react-spring": "^9.4.5",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
